requirements.txt
RPi.GPIO==0.7.0
paho-mqtt==1.5.0
loguru==0.4.1